==> 05-autogluon-tabular-simple.ipynb <==
## AutoGluon 활용

- tabular (housing price) : 앞에 autopilot과 동일한 데이터셋에 대해서 테스트



### 참고할 점
- autopilot 이 내부적으로 autogluon을 사용 함 : https://aws.amazon.com/ko/blogs/machine-learning/amazon-sagemaker-autopilot-is-up-to-eight-times-faster-with-new-ensemble-training-mode-powered-by-autogluon/
- autogluon 을 sagemaker 활용해서 학습 가능 : https://auto.gluon.ai/stable/tutorials/cloud_fit_deploy/cloud-aws-sagemaker-training.html#

# 설치가 필요한 경우 아래 주석을 제거해 주세요
!pip install -q autogluon

import pandas as pd
import numpy as np

from autogluon.tabular import TabularDataset, TabularPredictor

from sagemaker.session import Session

sagemaker_session = Session()
bucket = sagemaker_session.default_bucket()
print(bucket)

train_s3_path = f"s3://{bucket}/lowcode-sm/hp/train.csv"
train_df = pd.read_csv(train_s3_path)
train_df

test_s3_path = f"s3://{bucket}/lowcode-sm/hp/test.csv"
test_df = pd.read_csv(test_s3_path)
test_df

label = "SalePrice"

train_data = TabularDataset(train_s3_path)
test_data = TabularDataset(test_s3_path)


predictor = TabularPredictor(label=label).fit(train_data=train_data)


predictions = predictor.predict(test_data)

print(predictions)
